WIP: terraform: add lb terragrunt unit
This configures a load balancer with a set of SSL certificates managed by Azure.

Creating the certificate orders required registerin the
`Microsoft.CertificateRegistration` namespace in the subscription.

This is still work in progress, the certificate order doesn't have a
certificate yet, even though the domain validation token exists in DNS.

Maybe the correct way is
https://github.com/hashicorp/terraform-provider-azurerm/issues/24855
instead, or another additional resource. Documentation on this
unfortunately is a bit lacking...